+++ deep_translator/__main__.py
@@ -1,5 +1,4 @@
 
-import argparse
 from .cli import CLI
