+++ deep_translator/google_trans.py
@@ -4,7 +4,7 @@ google translator API
 
 from .constants import BASE_URLS, GOOGLE_LANGUAGES_TO_CODES, GOOGLE_LANGUAGES_SECONDARY_NAMES
 from .exceptions import TooManyRequests, LanguageNotSupportedException, TranslationNotFound, NotValidPayload, RequestError
-from .parent import BaseTranslator
+from .base import BaseTranslator
 from bs4 import BeautifulSoup
 import requests
 from time import sleep
